fixing bugs in every thing

+++ app/models/studentmanagedcourse.js
@@ -5,7 +5,7 @@ module.exports = function(sequelize, DataTypes) {
             IdStudent: DataTypes.INTEGER,
             IdCourse: DataTypes.INTEGER,
 			Name : DataTypes.STRING,
-			CreditPoints : DataTypes.INTEGER,
+			CreditPoints : DataTypes.DOUBLE,
 			CourseType: {
 				type:   DataTypes.ENUM,
 				values: ['mandatory','mandatory_choice','design_choice','special_projects','general_choice','free_choice','extra']
